+++ qmctorch/scf/calculator/__init__.py
@@ -1,5 +1,5 @@
-__all__ = ['CalculatorBase', 'CalculatorADF', 'CalculatorPySCF']
+__all__ = ['CalculatorBase', 'CalculatorADF', 'CalculatorADF2019', 'CalculatorPySCF']
 
 from .calculator_base import CalculatorBase
-from .adf import CalculatorADF
+from .adf import CalculatorADF, CalculatorADF2019
 from .pyscf import CalculatorPySCF