+++ qmctorch/scf/calculator/calculator_base.py
@@ -1,5 +1,3 @@
-from types import SimpleNamespace
-
 import numpy as np