+++ qmctorch/wavefunction/__init__.py
@@ -2,3 +2,11 @@ __all__ = ['WaveFunction', 'SlaterJastrow']
 
 from .wf_base import WaveFunction
 from .slater_jastrow import SlaterJastrow
+__all__ = ['WaveFunction', 'SlaterJastrow', 'SlaterOrbitalDependentJastrow']
+
+from .wf_base import WaveFunction
+from .slater_jastrow import SlaterJastrow
+# from .slater_combined_jastrow import SlaterManyBodyJastrow
+# from .slater_jastrow_backflow import SlaterJastrowBackFlow
+# from .slater_combined_jastrow_backflow import SlaterManyBodyJastrowBackflow
+from .slater_orbital_dependent_jastrow import SlaterOrbitalDependentJastrow