+++ tests/solver/test_h2_pyscf_stats.py
@@ -5,11 +5,11 @@ import torch
 import torch.optim as optim
 
 from qmctorch.sampler import Metropolis
-from qmctorch.solver import SolverSlaterJastrow
-from qmctorch.utils.plot_data import (plot_block, plot_blocking_energy,
-                                      plot_correlation_coefficient,
-                                      plot_integrated_autocorrelation_time,
-                                      plot_walkers_traj)
+from qmctorch.solver import Solver
+from qmctorch.utils import (plot_block, plot_blocking_energy,
+                            plot_correlation_coefficient,
+                            plot_integrated_autocorrelation_time,
+                            plot_walkers_traj)
 from qmctorch.scf import Molecule
 from qmctorch.wavefunction.jastrows.elec_elec import JastrowFactor, PadeJastrowKernel
 from qmctorch.wavefunction.slater_jastrow import SlaterJastrow
@@ -58,7 +58,7 @@ class TestH2Stat(unittest.TestCase):
         self.opt = optim.Adam(self.wf.parameters(), lr=0.01)
 
         # solver
-        self.solver = SolverSlaterJastrow(wf=self.wf, sampler=self.sampler,
+        self.solver = Solver(wf=self.wf, sampler=self.sampler,
                                           optimizer=self.opt)
 
     def test_sampling_traj(self):